+++ CHANGELOG.md
@@ -19,7 +19,8 @@
   
   ```let args = vec![1.to_es_value_facade(), 2.to_es_value_facade(), true.to_es_value_facade(), "sdf".to_string().to_es_value_facade()]```
 * added quickjs_utils::modules::new_module/add_module_export/set_module_export
-* added NativeModuleLoader to QuickJSRuntime to enable implementors to load native modules on-demand  
+* added NativeModuleLoader to QuickJSRuntime to enable implementors to load native modules on-demand
+* altered reflection to enable creation of JSValueRef without making it available in global scope (.install(ctx, false))
 
 # 0.1.0